Rollup merge of Gerd Stolpmann's iOS changes, excluding buildsystem ones.
Adaptions to the syntax of Apple as. This also includes: - workaround for the generation of jump tables in thumb mode - workaround for the generation of frametables - basic changes in arm.S - introduction of ldgaddr and ldladdr macros in arm.S which will be exploited later in the PIE patch - a little EABI fixup in arm.S: saving all FP regs where needed (EABI_APPLE is an EABI_HF variant)

Signed-off-by: Gerd Stolpmann <gerd@gerd-stolpmann.de>

On the iPhone simulator the machine context field has the two underscores.

Signed-off-by: Gerd Stolpmann <gerd@gerd-stolpmann.de>

Instructions for div and mod: - on ARMv7s there are machine instructions - IOS defines __divsi3 and __modsi3 instead of __aeabi_idiv and __aeabi_imod, resp

Signed-off-by: Gerd Stolpmann <gerd@gerd-stolpmann.de>

The apple assembler does not understand some of the pre-UAL mnemonics. Support both types of mnemonics, and select the right type dynamically.

Signed-off-by: Gerd Stolpmann <gerd@gerd-stolpmann.de>

Conflicts:
	asmcomp/arm/emit.mlp

for loading fp immediate constants, Apple as needs the fp as decimal number (sic)

Signed-off-by: Gerd Stolpmann <gerd@gerd-stolpmann.de>

implementing the ext calling conventions for Apple EABI. This is a little bit tricky, as these conventions do not directly fit to the framework (which doesn't support register pairs). The workaround is that we pretend that the first two float args are passed via d0 and d1, and that the single float result comes back in d0. This is not the truth, but fits into the framework. In reality, the float args go into r0/r1, r1/r2 or r2/r3, and are mixed with any int or pointer args. The float result is in r0/r1. When we emit the call instruction, we also emit additional code that moves the floats from d0 and d1 to the register pairs (and potentially also moves the integers away that are already there). After returning from the call, the float, if any, is loaded into d0. The price of this workaround: up to three additional moves per extcall.

Signed-off-by: Gerd Stolpmann <gerd@gerd-stolpmann.de>

Generate PIE code for Apple. This only affects symbol literals. Jumps can be resolved by ld (that's where PIE is simpler than PIC).

Signed-off-by: Gerd Stolpmann <gerd@gerd-stolpmann.de>

arm64 fix in instruction selection: the arm64 instruction set only supports unsigned immediates. Only for add/sub we can handle negative immediates by inverting the operation. (This is for all platforms, not only iOS.)

Signed-off-by: Gerd Stolpmann <gerd@gerd-stolpmann.de>

arm64 fixing assembler notation glitches: (1) index immediates need a hash mark according to ARM; (2) avoiding multi-statement C macros and using assembler macros instead

Signed-off-by: Gerd Stolpmann <gerd@gerd-stolpmann.de>

arm64: adapt to specific notational variants of apple assembler

Signed-off-by: Gerd Stolpmann <gerd@gerd-stolpmann.de>

Conflicts:
	asmcomp/arm64/emit.mlp

fix: the instructions lsls, lsrs, and asrs do not support an immediate arg of #0 in thumb mode. Replacing by a simple mov.

Signed-off-by: Gerd Stolpmann <gerd@gerd-stolpmann.de>

fix arm: using the correct UAL mnemonics for add/multiply operations

Signed-off-by: Gerd Stolpmann <gerd@gerd-stolpmann.de>

Define the new sub architecture for iOS: EABI_APPLE, ARMv7s, configurable support for PIE. PIC is now on if supported for the model.

Signed-off-by: Gerd Stolpmann <gerd@gerd-stolpmann.de>

Conflicts:
	asmcomp/arm/emit.mlp

iOS: the symbol __EABI__ is not defined; need to take this into account

arm: always using the modern UAL mnemonics